+++ src/main/java/org/truffleruby/platform/DarwinAArch64NativeConfiguration.java
@@ -39,7 +39,7 @@ package org.truffleruby.platform;
 
 import org.truffleruby.RubyContext;
 
-public class DarwinARM64NativeConfiguration extends DefaultNativeConfiguration {
+public class DarwinAArch64NativeConfiguration extends DefaultNativeConfiguration {
 
     public static void load(NativeConfiguration configuration, RubyContext context) {
         // Generated from tool/generate-native-config.rb on arm64-darwin20